backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
email-validator>=2.0.0
SQLAlchemy>=2.0.0
pytest>=7.4.0
requests>=2.31.0
PyJWT>=2.8.0
bcrypt>=4.1.3